expose FeatherSync
