+++ src/lib/icons/Feather.svelte
@@ -1,20 +1,8 @@
 <script lang="ts">
-	interface IconData {
-		contents: string;
-		attrs: {
-			xmlns: string;
-			width: number;
-			height: number;
-			viewBox: string;
-			fill: string;
-			stroke: string;
-			'stroke-width': number;
-			'stroke-linecap': 'inherit' | 'butt' | 'round' | 'square';
-			'stroke-linejoin': 'inherit' | 'round' | 'miter' | 'bevel';
-		};
-	}
+	import type { ComponentProps } from 'svelte';
+	import FeatherSync from './FeatherSync.svelte';
 
-	export let icon: IconData | Promise<IconData> | (() => Promise<IconData>);
+	export let icon: Promise<Pick<ComponentProps<FeatherSync>, 'contents' | 'attrs'>>;
 
 	export let label = '';
 	export let style = '';
@@ -23,36 +11,12 @@
 	export let flip: undefined | 'x' | 'y' = undefined;
 	export { className as class };
 	let className = '';
-
-	function scope<T>(fn: () => T) {
-		return fn();
-	}
 </script>
 
-{#await typeof icon === 'function' ? icon() : icon}
+{#await icon}
 	<slot name="loading" />
 {:then { attrs, contents }}
-	{@const { width, height } = scope(() => {
-		const { width: w, height: h } = attrs;
-		const ratio = Math.max(w, h) / 16 || 1;
-		return {
-			width: +scale * (w / ratio),
-			height: +scale * (h / ratio),
-		};
-	})}
-	{@const data = Object.assign(attrs, {
-		style,
-		width,
-		height,
-		'aria-label': label || null,
-		'stroke-width': girth,
-		role: label ? 'img' : 'presentation',
-		class: className || null,
-	})}
-
-	<svg {...data} style:transform={flip ? `scale${{ x: 'Y', y: 'X' }[flip]}(-1)` : ''}>
-		{@html contents}
-	</svg>
+	<FeatherSync {contents} {attrs} {label} {style} {girth} {scale} {flip} class={className} />
 {:catch error}
 	<slot name="error">
 		<pre>{JSON.stringify(error)}</pre>
