+++ src/lib/icons/FluentSync.svelte
